feat: 아이템 수정, 삭제 api 추가

+++ src/main/java/com/project/WebStore/item/repository/PointBoxItemRepository.java
@@ -5,7 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PointBoxItemRepository extends JpaRepository<PointBoxItemEntity, Integer> {
-
+public interface PointBoxItemRepository extends JpaRepository<PointBoxItemEntity, Long> {
   boolean existsByName(String name);
 }
